return 404 error code if the exam to delete do not exist

+++ exams-management-system/Controllers/ExamsController.cs
@@ -84,6 +84,12 @@ namespace exams_management_system.Controllers
         [HttpDelete("{id:guid}", Name = "DeleteExam")]
         public async Task<IActionResult> DeleteExam(Guid id)
         {
+            var exam = await this.examService.FindById(id);
+            if (exam == null)
+            {
+                return StatusCode(StatusCodes.Status404NotFound);
+            }
+            
             if (await this.examService.Delete(id))
             {
                 return Ok("Exam deleted");
